Migration of OtherSANs Parsing Call to SDK helper from pki-issuer
